BACKPORT: ARM: 7456/1: ptrace: provide separate functions for tracing syscall {entry,exit}
The syscall_trace on ARM takes a `why' parameter to indicate whether or
not we are entering or exiting a system call. This can be confusing for
people looking at the code since (a) it conflicts with the why register
alias in the entry assembly code and (b) it is not immediately clear
what it represents.

This patch splits up the syscall_trace function into separate wrappers
for syscall entry and exit, allowing the low-level syscall handling
code to branch to the appropriate function.

Reported-by: Al Viro <viro@zeniv.linux.org.uk>
Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>

Change-Id: Id327b2feb3729175d1bb9caaf6288e7197d23460
(backported from commit ad722541147e6e517a2077e3d944105e7bc4fa8e)
[ kyle: backport to krillin: Also changed the ret_from_fork
  syscall_trace to a syscall_trace_exit. ]
Signed-off-by: Kyle Fazzari <kyle@canonical.com>