UBUNTU: SAUCE: Enable squashfs for krillin.
This change is part of an effort to enable the installation
and running of .snap packages on the Ubuntu Phone. The Snappy
package format is soon moving to utilize a squashfs image per
package; installing such packages requires kernel support.

These config changes were copied from the trusty kernel, though
a few options weren't available and BCJ filter decoding was only
enabled for ARM. Still verified to work for the Snappy use-case.

Change-Id: I8c51c67c227ae9a1abf565b81b2d8e4816d6d352
Signed-off-by: Kyle Fazzari <kyle@canonical.com>