UBUNTU: SAUCE: Enable squashfs for vegetahd.
This change is part of an effort to enable the installation
and running of .snap packages on the Ubuntu Phone. The Snappy
package format is soon moving to utilize a squashfs image per
package; installing such packages requires kernel support.

These config changes were copied from the trusty kernel, though
a few options weren't available and BCJ filter decoding was only
enabled for ARM. Still verified to work for the Snappy use-case.

Change-Id: I34185a588f6d4461b11ca35365331bf303d24658
Signed-off-by: Kyle Fazzari <kyle@canonical.com>